modify and restructure the algorithm
