+++ src/dynamic_programming/mod.rs
@@ -2,6 +2,7 @@ mod coin_change;
 mod egg_dropping;
 mod fibonacci;
 mod fractional_knapsack;
+mod get_longest_palindrome;
 mod is_subsequence;
 mod knapsack;
 mod longest_common_subsequence;
@@ -28,6 +29,7 @@ pub use self::fibonacci::memoized_fibonacci;
 pub use self::fibonacci::nth_fibonacci_number_modulo_m;
 pub use self::fibonacci::recursive_fibonacci;
 pub use self::fractional_knapsack::fractional_knapsack;
+pub use self::get_longest_palindrome::get_longest_palindrome;
 pub use self::is_subsequence::is_subsequence;
 pub use self::knapsack::knapsack;
 pub use self::longest_common_subsequence::longest_common_subsequence;
