+++ src/string/mod.rs
@@ -4,7 +4,6 @@ mod autocomplete_using_trie;
 mod boyer_moore_search;
 mod burrows_wheeler_transform;
 mod duval_algorithm;
-mod get_longest_palindrome;
 mod hamming_distance;
 mod jaro_winkler_distance;
 mod knuth_morris_pratt;
@@ -29,7 +28,6 @@ pub use self::burrows_wheeler_transform::{
     burrows_wheeler_transform, inv_burrows_wheeler_transform,
 };
 pub use self::duval_algorithm::duval_algorithm;
-pub use self::get_longest_palindrome::get_longest_palindrome;
 pub use self::hamming_distance::hamming_distance;
 pub use self::jaro_winkler_distance::jaro_winkler_distance;
 pub use self::knuth_morris_pratt::knuth_morris_pratt;
